Update Explore event - Wicked family.tw (#224)
* Update Explore event - Wicked family.tw

Added ability to either trade for the young woman's freedom (without having sex) or buy her as a slave (after sex).
Also, fixed the event so that it subtracts food and wood from inventory when you give it to the old couple.  Makes sense to get some goodwill too.

* Add files via upload

New type of event for car breakdowns (to occur randomly based on vehicle's damage level)