Update Guest house.tw (#241)
* Update Guest house.tw

Added logic for sorting by amount of pregnancy, ascending and descending.
Please make sure to test that the arrow characters (&#11165; and &#11167;) are processed correctly.  It worked in the HTML, but I'm not sure if it needs any additional escaping when entered this way.

* Update Guest house.tw

It looks a lot better without the parentheses.